unit-tests: ExpiringMapTest integration-tests: com.dotmarketing.startup.runonce.Task220413IncreasePublishedPushedAssetIdColTest postman-tests: Apps.postman_collection,Scripting Resource.postman_collection

+++ .github/actions/run-postman-tests/src/postman.ts
@@ -195,7 +195,7 @@ const stopDotCMS = async () => {
     core.warning(`Could not stop gracefully DotCMS due to: ${err}`)
   }
 
-  if (!dotCmsProcess.killed) {
+  if (dotCmsProcess && !dotCmsProcess.killed) {
     killProcess(dotCmsProcess)
 
     await waitFor(20, 'DotCMS to stop')
